chore: split widgets on the smaller classes
